[QPL-663] Fix build warnings on gcc (part 3) (#485)
* [QPL-663][Fix] Trying more appropriate way to fix warnings on unused function parameters

* [QPL-663][Fix] Trying to fix avoid of warnings on unused parameters

* [QPL-663][Fix] Fixed lots of warnings on unused parameters

* [QPL-663][Fix] Fixed build

* [QPL-663][Fix] Fixed more warnings on unused parameters

* [QPL-663][Fix] Trying to fix warnings on casting between incompatible function types

* [QPL-663][Fix] Removed some unused variables

* [QPL-663][Fix] Temporaryly removed stack-usage flag

* [QPL-663][Fix] Removed some unused variables, fixed warnings on unused parameters

* [QPL-663][Fix] Fixed build

* [QPL-663][Refactoring] Renamed OWN_UNREFERENCED_PARAMETER macros

* [QPL-663][Fix] Testing implicit initialization warning fix

* [QPL-663][Fix] Another try to fix warning about implicit initialization

* [QPL-663][Fix] Fixing Windows build

* [QPL-663][Fix] Fixed new warnings on clang

+++ sources/middle-layer/analytics/input_stream.cpp
@@ -70,7 +70,7 @@ auto input_stream_t::unpack<analytic_pipeline::prle>(limited_buffer_t &output_bu
         // Situation when (prle_count_ == elements_left(), but current_source_size == 0) means the following:
         // The whole PRLE stream was unpacked, but there are not enough place in internal buffer to store all these values.
         // So all remaining repeating values will be stored in the next iteration, everything's OK, that's not an error.
-        if (prle_count_ != input_stream_t::elements_left()) {
+        if ((uint32_t)prle_count_ != input_stream_t::elements_left()) {
             return unpack_result_t(status_list::source_is_short_error);
         }
     }
